requirements.txt
playwright
beautifulsoup4
python-docx
python-dotenv
crewai
langchain-openai
notion-client